libc/rpc: bound the server table's index at both ends, and reject hostlen 0
xprt_register() indexes __svc_xports with xprt->xp_fd and guards only
from above:

	if (sock < FD_SETSIZE) {
		__svc_xports[sock] = xprt;

__xprt_do_unregister() reads it the same way. The floor is supposed to
be at the caller and is not there: svc_fd_create() takes the descriptor
from ITS caller and checks with assert(fd != -1), which does not reject
-2 and which bsd.debug.mk compiles out under MK_ASSERT_DEBUG=no. The
registration also happens before any validation -- makefd_xprt() sets
xp_fd and calls xprt_register() before the _getsockname() that would
notice a bad descriptor.

So: the kern_descrip.c idiom, one comparison for both ends.

		if ((unsigned int)sock < FD_SETSIZE) {

The `else if (sock == FD_SETSIZE)' arms are already exact and are left
alone -- the allocation really is FD_SETSIZE + 1 entries.

netname2host() in the same directory clamps with

	if (vallen > (hostlen - 1))
		vallen = hostlen - 1;
	(void) strncpy(hostname, val, vallen);

vallen is a pointer difference inside one string and cannot go negative
by itself, but hostlen - 1 is -1 when a caller passes zero, and
strncpy's third argument is size_t. There is no buffer a caller can
describe as length zero, so it joins the other malformed inputs the
function already returns 0 for. No in-tree caller; it is a public
netname2host(3) that keyserv and the NIS+ tools use.

Measured on lib/libc/rpc at HEAD: ONESIDED 3 -> 1, the survivor being
netname2host's own vallen, whose floor is the pointer difference and
stays invisible to a textual rule. Analyser 25 -> 25 with 56 of 56
translation units compiling both ways -- it never reported either of
these, which is why the tree runs more than one instrument.

Both markers revert-verified: exit 1 with either file reverted, exit 0
with both restored. Ledger regenerated for the new file lengths.

+++ hbsd/src/lib/libc/rpc/svc.c
@@ -110,7 +110,15 @@ xprt_register(SVCXPRT *xprt)
 		}
 		memset(__svc_xports, '\0', (FD_SETSIZE + 1) * sizeof(SVCXPRT *));
 	}
-	if (sock < FD_SETSIZE) {
+	/*
+	 * Bound both ends in one comparison: makefd_xprt() calls this
+	 * with whatever fd svc_fd_create() was handed, and that is only
+	 * assert(fd != -1) -- which does not reject -2, and is gone
+	 * entirely under NDEBUG.  A negative fd here is a write through
+	 * __svc_xports before the allocation.  The unsigned cast is the
+	 * same idiom kern_descrip.c uses for a descriptor.
+	 */
+	if ((unsigned int)sock < FD_SETSIZE) {
 		__svc_xports[sock] = xprt;
 		FD_SET(sock, &svc_fdset);
 		svc_maxfd = max(svc_maxfd, sock);
@@ -145,7 +153,8 @@ __xprt_do_unregister(SVCXPRT *xprt, bool_t dolock)
 
 	if (dolock)
 		rwlock_wrlock(&svc_fd_lock);
-	if ((sock < FD_SETSIZE) && (__svc_xports[sock] == xprt)) {
+	if (((unsigned int)sock < FD_SETSIZE) &&
+	    (__svc_xports[sock] == xprt)) {
 		__svc_xports[sock] = NULL;
 		FD_CLR(sock, &svc_fdset);
 		if (sock >= svc_maxfd) {
